src/main/java/com/sridhar/code/library/acompany/microsoft/Debug.java
package com.sridhar.code.library.acompany.microsoft;

public class Debug {

    public int solution(int[] S) {
        int max_sum = 0;
        int current_sum = 0;
        boolean positive = false;
        int n = S.length;
        for (int i = 0; i < n; ++i) {
            int item = S[i];
            if (item < 0) {
                if (max_sum < current_sum) {
                    max_sum = current_sum;
                    // this line is buggy.. most likely.
                    current_sum = 0;
                }
            } else {
                positive = true;
                current_sum += item;
            }
        }
        if (current_sum > max_sum) {
            max_sum = current_sum;
        }
        if (positive) {
            return max_sum;
        }
        return -1;
    }
}
